requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
streamlit
pandas
plotly
pdfplumber
python-docx
langchain
langgraph
openai
tiktoken
redis
sqlalchemy
psycopg2-binary
python-dotenv